Fix properties of type array<uuid> in migrations (#4114)
The issue is that the type array<uuid> was in an inconsistent
state. The shadow reflection schema uses array<uuid>, so it is always
in the reflected schema, but since it is created by the reflection
schema code and does not appear in the stdlib, it didn't appear in the
stdschema.

Fix this by adding any collection types that appear in the reflschema
into the stdschema.

Also to avoid related inconsistencies, skip loading objects that
appear in the stdschema when processing the json from the reflection
query. The query filters out qualified objects from the std schema but
not "global" ones, so collections and some other things were being
duplicated between _base_schema and _top_schema.

Fixes #4107.