Remove (mostly) the WHEN clause from access policies
I decided to go halfway with the compatibility break with old DDL: we
accept a WHEN in CREATE, and merge it with the expr, but we *don't*
accept modifications of WHEN in ALTER, which would require
persistently keeping track of whether a WHEN was present and handling
it in a bunch of places.

This should cover most existing DDL, I suspect.

Maybe it's only worth trying to deal with the compatability at all if
we do it fully, though.