Fix diamond pattern delegated constraints (#4123)
If we have an inheritence tree like V <: S, T <: R, with R
declaring a delegated constraint, we need to check both S and T
in the constraint trigger for V.

This requires a decent amount of rearranging the guts of trigger
generation.

Also fix dropping a property with such a constraint, which was the
originally filed issue here.

Fixes #4105.