Fix overly permissive cycle detection
Rework the tweak to cycle detection to continue to prohibit a child
class defining an access policy that refers to an ancestor, since that
would require generating a type filter for the ancestor that applies
policies to some but not all of its descendants!

test_edgeql_policies_cycle_02 is the relevant test we are trying to
prohibit, and test_edgeql_policies_cycle_03 is an obviously wrong case
that nevertheless passed in an earlier attempt at fixing this.