Prohibit access policy expressions with multi cardinality
I think it's semantically unclear what a multi cardinality value ought
to mean in a `deny` rule. There are two ways to think about what deny
rules should mean:
 1. Any object for which a deny rule returns `true` should be removed
    from the candidate set of objects
 2. We evaluate the following condition as a filter on each object
   `(allow_1 OR ... OR allow_n) AND NOT deny_1 ... AND NOT deny_m`

These definitions coincide if each rule has cardinality ONE, but they
differ if a rule can have cardinality AT_LEAST_ONE: if a rule returns
`{true, false}`, then definition 1 would exclude the object while
definition 2 would include it.

Currently the implementation follows definition 2, because it operates
based on boolean arithmetic; I think that if we had to choose that
definition 1 would be less surprising behavior, but that the best
option is to disallow multi cardinality, which lets our two ways of
thinking about the rules coincide.

We can always add it back in later if we want.