Fix a number of related access policy semantics issues (#4149)
1. Fix semantics of which policies apply *inside* of access policies

While evaluating a rule, the subject type *of the rule* and
*all of its descendants* may be accessed without restriction.
I think this is the only coherent thing.

Implementing this requires being substantially more intentional in the
compiler about when we want to use a type rewrite: previously we would
essentially always use a type rewrite, except when in the process of
compiling the type rewrite CTE itself, and that decision was made on
the pgsql side. There existed an `ignore_rewrites` mechanism, but it
was so-far only used to compile a conflict CTE for BaseObject when
doing an INSERT with explicit id.

We expand the use of `ignore_rewrites` to tag all references to a type
and its descendants while compiling an access policy defined on the
type. Since it needs to be based at the *source* of the type, we need
to compile each rule individually in order to compile them with the
proper exclusions.

2. DML policies also are evaluated without restriction on those types

Currently SELECT policies will be applied during the evaluation of DML
polices.  Change this for consistency, to avoid bizarre scenarios such
as being unable to update a SELECT-visible object of a type that only
has a single `allow all` rule, because it evaluates as true when
interpreted as a `select` rule but false when interpreted as an
`update read` rule.

3. Evaluate access policy rules in a detached context

Non-detached schema exprs cause a lot of trouble when they interact
with inheritance (see #4142). We discovered that issue too late to
deal with it properly for 2.0, since it will require a compatibility
break and a recovery strategy, but let's get it right for new features.

4. Adjust cycle detection

Rework the tweak to cycle detection to continue to prohibit a child
class defining an access policy that refers to an ancestor, since that
would require generating a type filter for the ancestor that applies
policies to some but not all of its descendants!

test_edgeql_policies_cycle_02 is the relevant test we are trying to
prohibit, and test_edgeql_policies_cycle_03 is an obviously wrong case
that nevertheless passed in an earlier attempt at fixing this.