Fix migration that removes policy using clause
I had made the (100% undocumented and untested) syntax for resetting a
policy USING clause be `RESET USING`. This was inconsistent with how
to reset every other single field that is set with `USING`, which is
`RESET EXPRESSION`, as well as the generic code in codegen that
generates it.  Switch it to `RESET EXPRESSION` and document it.

There was a schema test for removing a USING clause from a policy, but
that flow only ever operates on ASTs, and so the mismatch between the
codegen and the parser was undetecgted.

Fixes #4175.