Don't create constraints on derived views when adding a pointer to a type
If we have a derived schema view of a type (a global or an alias) and
try to add a pointer with a constraint on it to the base type, we will
also add the constraint to the derived, not real, view type. This is
a bad problem, because we do *not* include constraints and the like
when directly creating these views.

Make sure that we avoid doing this.

Also fix another bug I ran into when trying to test if this works in
the rebase case.

Fixes #4186.