Adapt setup.py/pyproject.toml to setuptools 54
The "editable mode" has been changed in >=setuptools-54 to no longer
call `setup.py develop` on which we've been relying so far.  And
although the `SETUPTOOLS_ENABLE_FEATURE=legacy-editable` escape hatch
exists, it probably wouldn't for long, so it's a good idea to switch
everything to be compatible with the brave new world of PEP 517/660.