Fix interaction between DETACHED and aliases/globals (#4321)
Referring to an alias in a detached block can result in the `expr` not
being attached to the set, which leads to it simply referring to the
underlying set. The cause here is a skew between
ctx.env.schema_view_cache and ctx.view_sets.

Fix this by separating the (cached) original compilation of a view and
the (uncached) population of ctx fields.

Fixes #4258.