Don't include union types in link triggers
If the schema contains a union type that has a link to some type M
(because it appears in every union element), we have been putting that
union type in the link triggers. This is bad because it isn't real.

Fixes #4320.

If you encounter this issue, after upgrading to a version with this
patch, it can be fixed by doing a dump/restore or by adding a new link
to `M` to the schema (and then deleting it, I guess).

It might be worth calling that out in the changelog.