Implement a patch application mechanism (#4355)
Include, in the source, a list of patches to apply, which can be either SQL or EdgeQL.

Then, in each database, track how many patches have been applied.

On server start, check if the system database has any unapplied patches. If so,
compile all the patches and apply them to every database that is missing them,
including the template database. Update the cached stdschema in the system database,
since that's the only one that matters. (Updating the stdschema is pretty marginal for
fixes to the bodies of functions, but would be important if we wanted to change a volatility,
for example.)

Also fix ALTER of a SQL function, which we'll need for most of these in practice.

Closes #4308.