Fix signalctl for Python 3.11
Python 3.11 deprecated the `msg` param in `Future.cancel()`, we can no
longer depend on the `msg` to mark the reason of cancellations. That
means, it is impossible to match recorded signals to CancelledErrors and
replace them with SignalErrors.

The solution here is to raise an ExceptionGroup containing all signal
errors, and its `__cause__` will be set to the cancellation error chain
displaying where the signal hit the user code. User error will be merged
into the same ExceptionGroup with signal errors, so be prepared to use
the `except*` grammar to handle exceptions.

One exception is CancelledError itself. If `SignalController.wait_for()`
is cancelled from outside (and no signals captured after that), a naked
CancelledError will still be raised. This made it possible to chain
`SignalController.wait_for()` to `asyncio.wait_for()` for timeouts.
Similarly, a group of signal errors can still be found in the __cause__.