Give better debug spew when a migration can't be resolved
Do this by inlining schemas_are_equal into the describe as json flow
and dumping the schema if they schemas aren't equal.

Then, delete schemas_are_equal because it only had one call site, and
it is a heavyweight enough operation that there's no reason to make it
convenient to add another.