Properly quote numeric names when in codegen (#4377)
Fixes #4344, in which dump wouldn't work in a schema with such a name.
Anybody who actually has such a schema is still in for a world of
pain, since the migration log won't dump properly. Fortunately, this
bug also *prevented* using the migration system to create types with
number names, so probably nobody is in this situation.