Fix potentially missed constraints on DML (#4410)
We use updates to a _dml_dummy table to force scans of constraint
checking CTEs. Unfortunately, if postgres analyzes the tables and
realizes that _dml_dummy is empty, this can cause all of these checks
to be optimized away.

Populate _dml_dummy, and run an analyze during template db creation to
catch this sort of issue.