Invalidate prior 2.x schema storage functions
Because of a bug fixed in #4401, the schema storage queries could
contain an EXISTS in a WHERE clause for .id constraint enforcement
that forced a table scan of every table in the database.

This caused pretty drastic performance problems. The fix has been
cherry-picked into 2.x, but existing databases will have the slow
versions already cached.

Invalidate the cache by adding `_v2` to the cache key.

This mechanism was chosen because it was much easier than writing a
PL/pgSQL patch that would delete all of the `__rh_*` functions, and
the extra cost of having a few orphaned functions sitting around is
not a big deal.