Fix creating an enum and a variadic function that references at the same time
The issue here is that function name mangling currently uses the *id*
of collection types. Collection type ids are deterministic and depend
on the ids of their components, but we need to generate the name
before the enum is created.

Just use the name.

This can't be cherry-picked to 2.0 because it changes the names of
lots of objects.

Fixes #4213.