Tweaks to speed up array indexing
While I was testing out my tuple argument stuff, it was *really* slow,
and the main culprit seemed to be array indexing.

One test case with maybe about a million elements total took long
enough to run that I gave up, so I went to investigate.

Just eliminating the "strict" on most of the array functions and their
helpers was enough to get the test runtime down to a managable 7s, but
eliminating the call to _int_or_null brought it down to 1.2s. I did
some experimentation, and merging the _int_or_null check into
_normalize_array_index also was able to bring it down to about
1.2s. (I'm not totally sure why it mattered so much.)

This meant I needed to adjust usage for the slice cases, so I pushed
the _int_bound into the call to _normalize_array_index.

I'm putting this up now in case you have any thoughts on the best way
to do this, but I think that I've probably messed the overflow edge
case handling and need to think about it a bit harder. Maybe we want a
separate `_normalize_array_slice_index` that does integrated slice
index normalizing and clamping?