Catch when POPULATE MIGRATION generates incorrect DDL
We currently can fail to catch if POPULATE MIGRATION generates DDL
that fails to produce the target schema (for example, #4483). This can
be caused both by issues with the DDL and with issues in producing an
incorrect target schema.

We already do appropriate correctness checks for the incremental
`DESCRIBE MIGRATION AS JSON` flow, but `POPULATE MIGRATION` is used to
initialize an empty database, and so it pretty key.

Make sure to catch this issue and produce an error, since it can lead
to very sticky database states.