Support passing tuples as arguments to queries
The key technique is to decompose a tuple parameter into multiple
parameters, one for each scalar component. Arrays of tuples are
essentially transposed into a tuple of arrays, and then decomposed.
Nested arrays are handled by flattening the array and adding an
`array<int32>` parameter that stores the start/end of each subarray.
Details are in the doc string for `tuple_args.py`.

In terms of data flow, the client sends properly encoded tuples, which
we encode in the server. We generate edgeql to convert the encoded data
into the correct types, and insert that into CTEs in the query.

Fixes #2326.