Make index names in postgres include the id of the index (#4490)
Currently index names include the *name* of the index, plus some other
stuff. The result of this is that they are long enough to need to get
hashed. Just make it based on the id, like we do with most things.

The upshot of this is that it will make it much easier for EXPLAIN to
figure out which index is being discussed.