In reordered SQL codegen, don't indent the bodies of UNION
We sometimes produce a big union, and it usually seems preferable to
have it formatted in a flat manner, instead of deeply indenting the
first components.

It might be better to understand nesting structure better and only
skip indenting when it is a basic left associative union, and not if
there is more interesting structure?