Make it possible to specify and toggle server readiness state
To distinguish between a "live" server and a "ready" server, add a new
`/server/state/alive` API to check for server liveness specifically.

The `/server/state/ready` check now depends not just on system liveness,
but also on the internal readiness state flag, the initial value of which
can be specified with `--initial-readiness-state={ready|not_ready}` or
its equivalent environment variable `EDGEDB_SERVER_INITIAL_READINESS_STATE`.

The readiness state can be set to `ready` via a `SIGUSR1` signal to the
main server process and to `not_ready` via a `SIGUSR2` signal.