Make ast.visitor.find_children more convenient and better typed
Every call to ast.visitor.find_children passes a predicate that calls
`isinstance`, and many of them pass a predicate that *only* calls
isinstance.

Take advantage of this pattern by adding a required type argument that
is now used to test the type of nodes before calling the
(now-optional) predicate function.

Then we can use that type argument to give precise types to the
predicate function and the output.

Drop the behavior where terminate_early would return an Optional[T]
instead of a List[T] since it is annoying to type (requires overloads)
and almost all the call sites are really just testing for existence
anyway.