Reimplement readiness state in terms of a file
I've changed my mind re the original design in this PR.  Specifying the
initial readiness state in the command line directly is problematic,
because if that is part of an orchestrator task definition, then the
server might be put into `not_ready` state inadvertently when a task is
restarted for any reason.  Thus, replace `--initial-readiness-state`
with a monitored `--readiness-state-file`.  Changing files in containers
should be as easy as sending a signal.