Don't lose type from inhviews when rebasing
Currently update_base_inhviews_on_rebase passes exclude_children to
alter_inhview when dropping a base, which results in a rebased type
being omitted from the inhview for any type that had been an ancestor
*before* the rebase. This might include the dropped base (if another
level was added in between) and *will* include std::Object and
std::BaseObject.

Fix this by not passing exclude_children, which wasn't needed.

While I was there, I also made it much more sparing in what types it
runs alter_inhview on. When changing a type's base, it will now redo
the inhviews for Object and BaseObject zero times instead of twice.

(The first added test tests the previously buggy behavior, the second
test worked before but I couldn't find an existing test for it.)