Fix SET REQUIRED on newly created properties with alias subtypes
`pgsql/delta` relies on `from_alias` in aux_data to determine whether
to perform various operations on pointers, so make sure it gets set on
newly created pointers on aliases.

Fixes #4382.

As a follow up I think that it might be an improvement to see if we
can ditch the whole `aux_data` mechanism.