Fix correlation issue related to factoring_allowlist (#4525)
When trying to reproduce looking at #4496, I got bogus "correlated
set" errors when running with typename injection on. The root of the
issue was that something related to the typename injection was causing
there to be an extra fence in between where factoring_fence was set
for an INSERT and the allegedly parent fence where the current
iterators were placed in the factoring_allowlist.

I fixed this by reworking things such that factoring_allowlist applies
to a factoring_fence on the *same* node, not on child nodes.