Fix casting empty sets to standard library enum types
When making up path_ids for empty sets, we need to keep track of
whether the actual type came from a standard library module, so that
we can propertly figure out whether it is in edgedbstd or edgedbpub.

The issue where this first came up was when casting the result of
json_set: an optional wrapper is being inserted for the
`empty_treatment` argument in this cast case. (That seems like its own
performance bug, too.)

Fixes #4530.