Fix arrays of ranges when used as argument types (#4543)
The backend OID computation for ranges introduced in #4463 doesn't
work properly, because it always instead grabs the type of an *array*
of the element type, not the type of the range. Pass the kind of type
being resolved as an additional argument to the OID resolution function,
so that we can distinguish properly.

We then also need to do the range OID resolution based on our
type_to_range_name_map, instead of based on the rngsubtype field,
since date_t uses daterange instead of a dedicated range.