Fix several issues that manifest when using GROUP BY (#4549)
It is common to hit materialization bugs when using GROUP BY. This fixes a
cluster of them.

1. When looking for a materialized pointer inside a source rvar
grabbing a pointer, discard any namespaces that are not on the source
path. unpack_rvar won't put those namespaces on the path, so we can't
either.
2. Explicitly track the binding scope for visibility materialization,
and use that when checking for visibility.

Fixes #3951. Fixes #4240. Fixes #4481.

Issues #3951 and #4240 depend on fix 1, #4481 depends on fix 2. Fix 2
also fixed two existing xerrored tests and turns one of the worst
remaining miscompiles into an ISE.