Add a `--use-db-cache` argument to `edb test`
Add a flag to tell `edb test` to attempt to maintain a cache of
initialized test databases.

This is unsound, of course--but no more unsound that
EDGEDB_DEBUG_BOOTSTRAP_CACHE_YOLO=1 is.

On my machine, this gets bootstrap time for my "fasttests" collection
of tests to 5s, down from 50s.

The change is most noticable when just running single tests: the time
to run a single select test is now 8s instead of 27s.