Make patches able to make modifications to schema layout
To do this, after applying any script component, we need to rerun
generate_structure, run any SQL it generates, and update all the
cached values it produces.

This also requires upgrading the regular edgeql script application
path to update reflschema and stdschema in sync and making
generate_structure able to run on an existing reflschema.