Fix bug if server crashes while applying object-creating patch (#4571)
When applying a patch, we compile the patch once, apply it to every
user and template db, and *then* apply it to the system db, which
serves to keep everything in sync.

However, if the server crashes after applying a patch to a user db,
but before it is applied to the system db, the patch will get
recompiled before we can apply it to the system db.

Currently, if a patch creates a new object (for example a function, or
a new object type in the introspection schema), this recompilation
will cause the object to have a different id in the system db and the
user db. Since the system db provides the std schema, this will cause
breakage whenever trying to use the object.

Fix this by recording the patches in the database and replaying them
from there when available, instead of recompiling.

No current patches we have shipped in our 2.x branch are vulnerable to
this problem, but some patches we probably will ship in the future
would be.

Tested by creating a patch that adds a function and causing the server
to crash before running the patch on the system db.