Make patches able to make modifications to schema layout (#4573)
To do this, after applying any script component, we need to rerun
generate_structure, run any SQL it generates, and update all the
cached values it produces.

This also requires upgrading the regular edgeql script application
path to update reflschema and stdschema in sync and making
generate_structure able to run on an existing reflschema.

To make sure that this doesn't totally break downgrade, add
version keys to many of the cached values in instdata. Also
make the upgrade CI tests do a downgrade and run a simple
query there.