Fix collection type using DDL and set global in notebook protocol
The collection type in DDL fixes some issues we've had with updating
the tutorial to use the latest edgedb, where a function taking
variadic args failed.

The approach to fixing it was to use the execute module instead of
trying to handle things directly.

Since this got us most of the way to having globals work in the
tutorial, I also got that working. The main snag there was the
capabilities system, which treats SET GLOBAL as a session
configuration command. This is true, but probably too strict: add a
new capabilities flag for SET GLOBAL, but make it implied by the
session config command, so as to not break protocol compatability.