Fix using WITH-bound DML from an UPDATE in an ELSE clause (#4641)
More generally, always process DML containing bindings at the point
the statement is compiled and not when the binding is first
referenced. Currently we handle this properly for SELECT but not for
other statements.

As a result of not handling that right, we were first compiling a
WITH-bound insert in the context of an UNLESS CONFLICT clause
containing an UPDATE and then trying to reuse it in the INSERT
that the UPDATE was hanging off of, which broke.

Fixes #4577.