Implement some mild hacks to allow testing SQL translator
Make 'DESCRIBE ALTER <string literal>' interpret string literal as raw
SQL source, rewrite it as necessary (TODO!) and return the textual
result.

I went with DESCRIBE ALTER as the (totally nonsensical) keyword for
this because I didn't want to have to adjust any keywords, it worked,
and the presence of ALTER prevents the query normalizer from
extracting the string constant.

Then also add a test class that executes SQL queries by getting the
pgaddr from dev-mode connection settings, running DESCRIBE ALTER to
get a rewritten query, and then using asyncpg to directly query the
underlying postgres instance.

The core virture of this approach is that we'll be able write test
cases, then swap out the implementation of SQLQueryTestCase and keep
using the same test cases.

Maybe we actually want to use psycopg instead, since it may be a
smoother path to transitioning the tests directly speak the pg
protocol to the edgedb server?

This is obviously a hack, but I think it is a mild enough one that it
is mergeable (unlike the previous version that invoked psql).