Fix putting a statement as the body of an access policy
Currently it isn't (by oversight) allowed. The particularly bad part,
though, is that while you can fix the parse by wrapping the statement
in parens, the DDL codegen will lose the parens again.

The easiest way to work around this is probably wrapping it in braces,
so it is a set literal.