Add wrappers for many more bytes functions
This exposes:
 * convert_from
 * convert_to
 * encode
 * decode
 * sha{224,256,384,512}

I kept all of their names, but prefixed the non-SHA ones with
`bytes_`.  I'm not confident in either of those decisions and so would
be happy to have discussion on that.

We'll need to add some tests and documentation, but I'm putting this
PR up in advance of that because I want to use a hash function to
generate deterministic "random" numbers for testing purposes.