Fix SET GLOBAL capabilities (#4688)
In #4605, in an attempt to fix SET GLOBAL for the notebook protocol, I
made SET GLOBAL use its own capabilities flag while attempting to
maintain backward compatability. I screwed up in both directions,
though, breaking JS client tests that wanted SET GLOBAL disabled
when SESSION_CONFIG was cleared, and (much worse) breaking SET GLOBAL
from the cli.

Roll back that mechanism and just hack the notebook into being able to
handle it.

It's kind of shocking to me that nobody has reported this. We should
try to get this out in a release ASAP. (Though maybe a better fast fix
would be on the CLI side?).