Clarify the `duration_get` function (#4655)
* Clarify the `duration_get` function

The behavior of `duration_get` is inconsistent. For duration elements
month or larger, the units are aggregated to return the final value. For
duration elements smaller than month, only the literal value expressed
in the duration for that unit is returned. Since this behavior could be
unexpected, we need additional documentation around it.

* Mark up "Gotchas" headings as rubrics

* Document that units <= hour are also totaled

I previously believed only month and larger were totaled

* Explain why size classes exist

* Clarify extractable units per type

* Add note about normalize/truncate in duration_get

* Remove errant quote

Co-authored-by: i0 <41456914+i0bs@users.noreply.github.com>

* Update duration_get one-liner

* Clarify `totalseconds` approximation

* Remove repetition

Co-authored-by: i0 <41456914+i0bs@users.noreply.github.com>