Update linkprop serialization in tests for new edgedb-python
Link props are now exposed on objects and doing indexing on an object
to get Link/LinkSet now raises deprecation warnings, so avoid doing
that and instead fetch the fields directly.

We read the fields out using dataclasses.fields, which seems to be
the only place the list of link properties are exposed.