Fix warning from unclosed subprocess transport in pool tests (#4726)
We needed to add some `close()` calls to the pool. I wanted to fix it
for the SimpleAdaptivePool as well, which meant testing that
also. Trying to do that I bumped into #4680 again, where we only scale
up based on the number of requests increasing over a window, which
doesn't match with this test that wants two workers at once. Here I
hacked around it for now.