Track the origin of schemas stored in an expression (#4731)
Track what object/field an expression is stored in and propagate this
information into context fields. It then should be possible to use
this in error reporting and explain.

We store the id and the field name; we store id instead of name
because once these values are reported to a client, id is more useful:
the client can't query based on the internal name.