[DRAFT] An initial take on EXPLAIN
Here is an initial take on EXPLAIN that hopefully is in a place that
it can be played around with for frontend tooling.

Mostly it returns postgres's JSON output for EXPLAIN, but rewrites
references to internal names into useful human readable names.

It also annotates subplans that actually touch tables with a list of
"Contexts" indicating corresponding edgeql source positions.
It is a list of contexts so that in queries like
`with U := User, U { name }`, we can indicate that the scan of `User`
originates from the reference to `User` most specifically and then
from the reference to `U` more generally. We also do this with
computed pointers, including computed pointers defined in the schema.

In order to support computed pointers defined in the schema, the
result also contains a 'Buffers' field that is a list of all the
source strings referenced in a context, starting with the actual query
text and including any expression from the schema used. When a schema
expression is referenced, we also return the object's id and the
name of the field.