Add more specific `DuplicateMigrationError` error
Motivation: it simplifies implementation of applying migrations by the
client bindings.

Downsides: we have to check whether hash of migration is consistent then
check whether migration exists and the last step is to check the parent
migration. Previously we have given feedback that parent is invalid
first. Which was occasionally helpful for editing migrations by hand.