Disallow casts to and literal references of enum types from indexes/constraints (#4754)
The casts that postgres generate are STABLE, though we consider them
to be IMMUTABLE. We'll need to introduce wrapper functions. Fixing it
in 2.x is probably more of a hassle than it is worth, so disable it
for now with the intention of fixing it in 3.0.

Mitigates #4592 by turning it from an ISE into a legible error, but
does not fix it.