Fix changing a link to non-computed and single at the same time
The issue is that the link table gets schedule for an inhview update
when it is created with RESET EXPRESSION, but then once it is made
single, it doesn't actually have a table anymore.

Fix by gating recreate_inhvew in apply_scheduled_inhview_updates with
a has_table check, like we do for alter_inhview.

Fixes #4716