Delete *all* of the types created by an alias
In some situations we create view types as part of aliases that
*don't* actually appear directly in the result type. Currently we fail
to clean up those types when the alias is delted or altered.

For now, fix this by searching for all such types (... by name
prefix) and deleting them.

For 3.x, we will want to fix this in a better way, probably one of:
 * Track the child views in the schema explicitly, instead of needing
   to search through the schema for them.
 * Ensure that we don't create include view types that don't actually
   appear in the output. This approach is probably better, but I'm not
   100% sure it works.

Fixes the second issue in #4769.