Remove mypy: ignore-errors from pgsql/delta.py
I haven't made any real effort to add more type annotations in, but
I've gotten it to the point where it does typecheck.

A decent amount *is* checked, though.

The big limitation in the typing support is that if we pull an adapted
`op` out of the context, we will only ever infer its *non-adapted*
type. If we want to call methods from pgsql/delta, we need to add an
assert that it is also some pgsql/delta command.