Fix link properties on inherited backlinks
Only include in the IR the pointers that actually could be used when
having a type intersected backlink. The tricky bit here (and what held
me up doing this in the past) is that we still must use a ptrcls with
the full opaque union of every pointer in the *pathid*, because
otherwise path factoring breaks

(consider `SELECT (User.<owner[IS Comment], User.<owner[IS Issue])`,
where `User.<owner` is a common path and so needs the same pathid in
each.)

This fixes an issue where we would uselessly include link tables that
couldn't be used for multi links, as well as an issue where we would
tried accessing the link property on the wrong link table.

Fixes #2571. Fixes #4704.