Fix some migrations with tricky constraint/computed interactions
The issue here is that the deletion of `id` and the deletion of a type
might get separated in the generated delta. While processing the delta
to produce an AST, expressions that reference the type might get
recompiled, and nasty things can happen since the id is missing.

Fix this by treating `id` as a "strong ref", which shouldn't ever have
its operations hoisted outside of the main operation on the enclosing
object.

Tweak some bits in ordering to not cause knock-on problems.

Fixes #4789