Implement extended query
The idea is to collect all extended-query messages from the inbound
buffer, build a list of backend actions and execute them all at once.
Each execution must end with a single SYNC in the backend, because the
pgcon will be released after execution. However, this doesn't mean we
always send a ReadyForQuery back to the frontend, because the client may
not send a SYNC yet. In this case, frontend implicit transaction is not
guaranteed.