Add documentation for interactive shell commands (#4823)
* Add documentation for interactive shell commands

* Add detail about command indexes in history

* Flip order of sentences

* Fix inline code markup

* Remove `-i` and add `[PATTERN]` to `\lm`/`\lr`

* Change section title

* Change `-I` to `-c`

Case-insensitive matching is now the default. `-c` is used to trigger
case-sensitive matching.

* Remove hyphen from `\describe schema`

* Clarify backslash command aliases

Duplicated the arguments so it's clear what parts of the alias are
needed. Some alises like `\ds` are commands and sub-commands bundled
(`\describe schema`) so the more verbose presentation is helpful to be
explicit about what the commands are.

* Documents migration backslash commands

* Make spacing consistent

* Rename heading

"Operations" is a little broad since it could be applied to many of the
commands that are not in this section.

* Document `\database create`

Co-authored-by: Paul Colomiets <paul@edgedb.com>