Get rid of the __std__derived__ hack
In #4530 I introduced a new __std_derived__ fake module for
distinguishing whether a pathid with a made up type name came from a
standard module or not, since that information is needed to generate
correct type names in the backend in some places (notably when dealing
with emptysets of standard library enum types).

This was a hack, and it made all our debug output uglier.  Instead,
track the original name of types in typerefs, even when the name has
been overridden.