Fix two issues with DML in free objects (#4902)
Fix an ISE where the volatility ref from the free object will
get referenced in the CTE for a for loop that contains DML
inside the free object. Fix this by not inserting volatility
refs for free objects, where they shouldn't be needed.
Fixes #4875.

Also fix an issue where DML in a non-exposed free object would not be
executed. Fix this by only allowing free object DML in exposed
positions.