[parser] Implement an inliner for PtrQuals
There are currently places where the definition of PtrQuals gets
inlined into the definition of another nonterminal, because parsing it
in a nicely factored way requires 2 lookahead tokens and we are LR(1).

Implement an inliner for `PtrQuals` that will expands out a reference
to `InlinedPtrQuals` into all of the appropriate rules. This can be
used for the definition of `PtrQuals` itself and anywhere else.