Implement manual error override configuration (#4974)
Implement a configuration field that causes the server to return an
error for all non-configuration commands.
See #4625 for details and motivation. Closes #4625.

The details here are a little different than proposed there.
There is one config property, `force_database_error`, that
accepts a json encoded string with the following format:
`false` means no error, and an object means to raise an error:
```
{
  "type": <name of exception class>,
  "message": <exception message>,
  "hint": <exception hint>,
  "details": <exception details>,
  "context": {
    "line": <line number>,
    "col": <column number>,
    "start": <start position>,
    "end": <end position>,
    "filename": <file name for error>
  }
}
```
All fields except for `type` are optional.