Optimize trivial WITH-bound `group` uses (#4978)
If the last alias in a WITH block is a GROUP query that is immediately
used in a shape or a for loop, inline it so that our other
optimizations apply.

This probably covers 95% of WITH-bound group use cases, despite being
a "hyper-local hack", as I said in #4243.

Fixes #4243. (Mostly.)