Fix a category of confusing scoping related bugs in access policies (#4994)
The edgeql->IR compiler has a notion of "weak namespaces" in which
scope tree nodes are tagged with namespaces, and those namespaces will
be dropped from a path if it refers to a node bound above that node.

The pg side doesn't understand any of this, though, so we have a
_rewrite_weak_namespaces that rewrites all affected path ids to
reference their targets directly.

Unfortunately we neglect to call it on our type rewrite expressions.
Fix that.

Also refactor `fini_expression` to try to make this category of bug
less likely by having more things driven by a list of expressions to
process.