Add WatchError
This is raised when `edgedb watch` have schema or migration error to
notify applications that schema is out of date.

It's generally bad to set the same error class as the original error
received by `edgedb watch` because various errors could be raised during
it's operation (i.e. duplicate key error when migration contains some
inserts). And all of them mean only one thing for the client: user
attention is needed, and client should not react to that error normally.