Avoid infinite recursion in some do-nothing intersection cases (#5007)
If a computed (or just appearing in a shape) link is referred to in a
clause with a type intersection that doesn't change the type, we
include it multiple times in our list of sets along the path, which
causes us to patch it up in an invalidly self referential way when we
handle computeds. Avoid doing that.

Fixes #4996.