Don't mangle cast error messages when the cast value contains a type name (#5008)
Currently if we do something like `<int64>'bigint'`, we will produce a
message like `invalid input syntax for type std::int64: "std::int64"`.
This is because we do a pretty simple-minded replacement of pg type
names with edgedb type names in this sort of error message.

Make it a bit less simple-minded: only do the replacement before the
first colon. This isn't particularly principled, but I don't think the
data is structured enough for principle.

Fixes #4999.