Implement compilation of triggers (#5088)
Closes #4936.

The interesting part of the compilation runs at the end of IR->SQL and
injects the triggers into CTEs. The overlay system is (somewhat
hackily) used to drive what objects are triggered on and (slightly
less hackily) to provide a view of the new database state.

I have one known bug at the moment, which is approximately that we
apply access policies *too much* in one case. I have a plan for fixing
it that doesn't change that much of the big picture, and I wanted to
get something up for review.