Fix a ISE when using assert_exists and linkprops using query builder
The core issue here was that the path id being used to unpack
materialized data had the wrong typeref. The root cause of this was
that in some situations we change the 'target' of a Pointer after a
PointerRef conversion of it is already cached.

Invalidate the cache entry when changing target.

Fixes #4961