Do not store __type__ in the backing database
Every edgedb object has a `__type__` link, but its value is determined
uniquely by the type, and so we can save a uuid column per object in
the database by not actually storing it.

Inject the `__type__` values explicitly into the inheritance views we
generate, and inject the values in the compiler when we operate
directly on tables.

Fixes #4637.