Eliminate ptr_info arg to get_path_output and similar
Instead of figuring out the column name for the actual pointer
(which could be a subtype of the one in the path id) in get_path_var
/get_rvar_path_var and threading it through, track typeref in Relation
so that we can find the actual ptrref when we get pointer path outputs
from Relations.

This allows us to do the correct column determination at a place that
makes more sense (_get_rel_path_output), and also gives
_get_rel_path_output access to the actual type information, which I'm
going to want in a follow-up PR to remove `__type__` from database
tables (#4637).