Try to print readable names in our "Reordered SQL" debug output
We do this in a pretty naïve way, which is to simply look for uuids
with a regexp and try to look up an object with that id. This doesn't
really work, but it's debug code and so it works great. We do
something similar in the pending EXPLAIN branch.

We only turn this on for when we dump "Reordered SQL" (enabled with
`EDGEDB_DEBUG_EDGEQL_COMPILE_SQL_REORDERED_TEXT=1`), on the theory
that the real SQL output ought to be something you can paste into psql
and have it work.